% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ad_drug_class_colors}
\alias{ad_drug_class_colors}
\title{Drug-class colour palette for chord diagrams}
\format{
A named character vector of length 7.
}
\usage{
ad_drug_class_colors
}
\description{
A named character vector mapping each of the 7 AD drug classes to a
distinct hex colour, used for sector bands in \code{\link[=plot_chord_dgi]{plot_chord_dgi()}}.
}
\examples{
data(ad_drug_class_colors)
ad_drug_class_colors["JAK inhibitor"]
}
\keyword{datasets}
