% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{dgi_edge_palette}
\alias{dgi_edge_palette}
\title{Interaction-type colour and line-type palette}
\format{
A tibble with 14 rows and 5 columns:
\describe{
\item{interaction_type}{Interaction type label in UPPER CASE (character).}
\item{edge_color}{Hex colour for network edges / links (character).}
\item{sector_color}{Hex colour for chord-diagram sector band (character).}
\item{link_color}{Hex colour for chord-diagram link ribbon (character).}
\item{lty}{Line type integer for chord links (integer).}
}
}
\usage{
dgi_edge_palette
}
\description{
A tibble mapping DGIdb interaction-type labels (upper-case) to hex colour
codes and line types used consistently across all \pkg{Repurp} plot functions.
}
\examples{
data(dgi_edge_palette)
dgi_edge_palette
}
\keyword{datasets}
