% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dgi_interaction_styles}
\alias{dgi_interaction_styles}
\title{Full interaction-style table for chord and sector colouring}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.
}
\usage{
dgi_interaction_styles
}
\description{
A tibble with columns \code{interaction_type}, \code{sector_color}, \code{link_color}, and \code{lty}.
}
\keyword{datasets}
