% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{dgi_node_colors}
\alias{dgi_node_colors}
\title{Node colour palette keyed by node type}
\format{
An object of class \code{character} of length 3.
}
\usage{
dgi_node_colors
}
\description{
Node colour palette keyed by node type
}
\examples{
dgi_node_colors["Gene"]
}
\keyword{datasets}
