% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chord.R
\name{repurp_class_colors}
\alias{repurp_class_colors}
\title{Default drug-class colour palette for chord diagrams}
\usage{
repurp_class_colors()
}
\value{
A named character vector.
}
\description{
Default drug-class colour palette for chord diagrams
}
